+++ tests/Diagnostics.test.cpp
@@ -1,0 +1,56 @@
+#include "doctest.h"
+#include "Fixture.h"
+#include "TempDir.h"
+#include "Platform/RobloxPlatform.hpp"
+
+TEST_SUITE_BEGIN("Diagnostics");
+
+TEST_CASE_FIXTURE(Fixture, "document_diagnostics_sends_information_for_required_modules")
+{
+    client->capabilities.textDocument = lsp::TextDocumentClientCapabilities{};
+    client->capabilities.textDocument->diagnostic = lsp::DiagnosticClientCapabilities{};
+    client->capabilities.textDocument->diagnostic->relatedDocumentSupport = true;
+
+    // Don't show diagnostic for game indexing
+    loadDefinition("declare game: any");
+
+    registerDocumentForVirtualPath(newDocument("required.luau", R"(
+        --!strict
+        local x: string = 1
+        return {}
+    )"), "game/Testing/Required");
+    auto document = newDocument("main.luau", R"(
+        --!strict
+        require(game.Testing.Required)
+    )");
+
+    auto diagnostics = workspace.documentDiagnostics(lsp::DocumentDiagnosticParams{{document}});
+    CHECK_EQ(diagnostics.items.size(), 0);
+    CHECK_EQ(diagnostics.relatedDocuments.size(), 1);
+}
+
+TEST_CASE_FIXTURE(Fixture, "document_diagnostics_does_not_send_information_for_required_modules_if_related_document_support_is_disabled")
+{
+    client->capabilities.textDocument = lsp::TextDocumentClientCapabilities{};
+    client->capabilities.textDocument->diagnostic = lsp::DiagnosticClientCapabilities{};
+    client->capabilities.textDocument->diagnostic->relatedDocumentSupport = false;
+
+    // Don't show diagnostic for game indexing
+    loadDefinition("declare game: any");
+
+    registerDocumentForVirtualPath(newDocument("required.luau", R"(
+        --!strict
+        local x: string = 1
+        return {}
+    )"), "game/Testing/Required");
+    auto document = newDocument("main.luau", R"(
+        --!strict
+        require(game.Testing.Required)
+    )");
+
+    auto diagnostics = workspace.documentDiagnostics(lsp::DocumentDiagnosticParams{{document}});
+    CHECK_EQ(diagnostics.items.size(), 0);
+    CHECK_EQ(diagnostics.relatedDocuments.size(), 0);
+}
+
+TEST_SUITE_END();