+++ tests/Fixture.h
@@ -39,6 +39,7 @@ struct Fixture
     ~Fixture();
 
     Uri newDocument(const std::string& name, const std::string& source);
+    void registerDocumentForVirtualPath(const Uri& uri, const Luau::ModuleName& virtualPath);
 
     Luau::AstStatBlock* parse(const std::string& source, const Luau::ParseOptions& parseOptions = {});
     Luau::LoadDefinitionFileResult loadDefinition(const std::string& source, bool forAutocomplete = false);